Use rustls instead of native openssl
This change makes building a statically linked executable later on a lot
easier, as everything needs to be build against musl for this. Building
openssl against musl is possible, but quite a pain. Therefore just using
the rust internal implementation makes things a lot smoother, and
shouldn't be of any further implication for our use case