Implemented compression of in memory image storage
The images fetched and evaluated are now stored as compressed PNGs
within memory. This should ease the memory usage of the application.

The update performance has been degraded a little bit because of this,
as all images have to be recompressed after retrieval. The composition
retrieval time however only changed negligable. Therefore this change
seems to be a good compromise between speed and memory consumption.